+++ domestic/models.py
@@ -318,6 +318,12 @@ class GreatDomesticHomePage(
         blank=True,
     )
 
+    def get(self, request, *args, **kwargs):
+        redirector = helpers.GeoLocationRedirector(self.request)
+        if redirector.should_redirect:
+            return redirector.get_response()
+        return super().get(request, *args, **kwargs)
+
     def _get_industry_tag_usage_counts(self, industry_tag):
         return industry_tag.countryguidepage_set.all().live().count()