+++ CHANGELOG.md
@@ -45,7 +45,7 @@ GP2-1849 - Feature flag magna links in logged out mobile header
 - GP2-1630 - add more data
 
 ### Implemented enhancements
-
+- GP2-1648 - Port Geolocation redirects from Bau UI
 - GP2-1852 - WTE responsive format
 - GP2-1849 - Logged out mobile header
 - GP2-1784 - Stop breadcrumbs showing for Structural Page